requirements.txt
fastapi
uvicorn
aiomysql
sqlalchemy
python-dotenv
